Massive refactor python pipeline to be 1-1 with core library
