+++ python/kiss_icp/pipeline/__init__.py
@@ -20,25 +20,4 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-import numpy as np
-
-from kiss_icp.config import KISSConfig
-from kiss_icp.pybind import kiss_icp_pybind
-
-
-def get_preprocessor(config: KISSConfig):
-    return Preprocessor(config)
-
-
-class Preprocessor:
-    def __init__(self, config: KISSConfig):
-        self.config = config
-
-    def __call__(self, frame: np.ndarray):
-        return np.asarray(
-            kiss_icp_pybind._preprocess(
-                kiss_icp_pybind._Vector3dVector(frame),
-                self.config.data.max_range,
-                self.config.data.min_range,
-            )
-        )
+from .kiss_icp import KissICP
